Refactor notification processor Add notification mute period by processor
