--- src/main/java/org/craftercms/deployer/impl/processors/MailNotificationProcessor.java
@@ -1,364 +1,0 @@
-/*
- * Copyright (C) 2007-2022 Crafter Software Corporation. All Rights Reserved.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 3 as published by
- * the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package org.craftercms.deployer.impl.processors;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.apache.commons.configuration2.Configuration;
-import org.apache.commons.io.FileUtils;
-import org.apache.commons.lang3.StringUtils;
-import org.craftercms.commons.config.ConfigurationException;
-import org.craftercms.commons.mail.Email;
-import org.craftercms.commons.mail.EmailFactory;
-import org.craftercms.deployer.api.ChangeSet;
-import org.craftercms.deployer.api.Deployment;
-import org.craftercms.deployer.api.ProcessorExecution;
-import org.craftercms.deployer.api.exceptions.DeployerException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.File;
-import java.io.IOException;
-import java.net.InetAddress;
-import java.net.UnknownHostException;
-import java.time.format.DateTimeFormatter;
-import java.util.*;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-import static org.craftercms.commons.config.ConfigUtils.*;
-
-/**
- * Post processor that sends an email notification with the result of a deployment, whenever a deployment fails or files where processed.
- * The output file is attached if it's available. A processor instance can be configured with the following YAML properties:
- *
- * <ul>
- *     <li><strong>templateName:</strong> The name of the Freemarker template used for email creation.</li>
- *     <li><strong>from:</strong> The value of the From field in the emails.</li>
- *     <li><strong>to:</strong> The value of the To field in the emails.</li>
- *     <li><strong>subject:</strong> The value of the Subject field in the emails.</li>
- *     <li><strong>html:</strong> Whether the emails are HTML.</li>
- *     <li><strong>serverName:</strong> The hostname of the email server.</li>
- *     <li><strong>dateTimePattern:</strong> The date time pattern to use when specifying a date in the email.</li>
- * </ul>
- *
- * @author avasquez
- */
-public class MailNotificationProcessor extends AbstractPostDeploymentProcessor {
-
-    private static final Logger logger = LoggerFactory.getLogger(MailNotificationProcessor.class);
-
-    /**
-     * Status conditions used to control whe the notifications should be sent.
-     * @author joseross
-     */
-    public enum StatusCondition {
-        /**
-         * Notifications will be sent for all deployments.
-         */
-        ON_ANY_STATUS,
-
-        /**
-         * Notifications will be sent for deployments in which at least one processor has failed.
-         */
-        ON_ANY_FAILURE,
-
-        /**
-         * Notifications will be sent for deployments in which the general status indicates failure.
-         */
-        ON_TOTAL_FAILURE
-    }
-
-    public static final String TEMPLATE_NAME_CONFIG_KEY = "templateName";
-    public static final String FROM_CONFIG_KEY = "from";
-    public static final String TO_CONFIG_KEY = "to";
-    public static final String SUBJECT_CONFIG_KEY = "subject";
-    public static final String HTML_CONFIG_KEY = "html";
-    public static final String SERVER_NAME_CONFIG_KEY = "serverName";
-    public static final String STATUS_CONDITION_KEY = "status";
-    public static final String DATETIME_PATTERN_CONFIG_KEY = "dateTimePattern";
-
-    public static final String SERVER_NAME_MODEL_KEY = "serverName";
-    public static final String TARGET_ID_MODEL_KEY = "targetId";
-    public static final String START_MODEL_KEY = "start";
-    public static final String END_MODEL_KEY = "end";
-    public static final String STATUS_MODEL_KEY = "status";
-    public static final String OUTPUT_ATTACHED_MODEL_KEY = "outputAttached";
-    public static final String PROCESSOR_MATCH_PATTERNS_CONFIG_KEY = "failedProcessors";
-    public static final String MUTE_PERIOD_MINUTES_CONFIG_KEY = "mutePeriodMinutes";
-    public static final String LAST_DATETIME_FILE_SUFFIX_CONFIG_KEY = "lastDateFilenameSuffix";
-
-    protected String defaultTemplateName;
-    protected String defaultFrom;
-    protected String defaultSubject;
-    protected boolean defaultHtml;
-    protected String defaultStatusCondition;
-    protected String defaultDateTimePattern;
-    protected String defaultLastDateFilenameSuffix;
-    protected String lastNotificationDateDir;
-    protected EmailFactory emailFactory;
-    protected ObjectMapper objectMapper;
-
-    // Config properties (populated on init)
-
-    protected String templateName;
-    protected String from;
-    protected String[] to;
-    protected String subject;
-    protected boolean html;
-    protected String serverName;
-    protected StatusCondition statusCondition;
-    protected Pattern failedProcessorsPattern;
-    protected int mutePeriodMinutes;
-    private String lastDateFilenameSuffix;
-    protected DateTimeFormatter dateTimeFormatter;
-
-    /**
-     * Sets the default name of the Freemarker template used for email creation.
-     */
-    public void setDefaultTemplateName(String defaultTemplateName) {
-        this.defaultTemplateName = defaultTemplateName;
-    }
-
-    /**
-     * Sets the default value of the From field in the emails.
-     */
-    public void setDefaultFrom(String defaultFrom) {
-        this.defaultFrom = defaultFrom;
-    }
-
-    /**
-     * Sets the default value of the Subject field in the emails.
-     */
-    public void setDefaultSubject(String defaultSubject) {
-        this.defaultSubject = defaultSubject;
-    }
-
-    /**
-     * Sets whether the emails are HTML by default.
-     */
-    public void setDefaultHtml(boolean defaultHtml) {
-        this.defaultHtml = defaultHtml;
-    }
-
-    /** Sets the default condition to send emails.
-     */
-    public void setDefaultStatusCondition(final String defaultStatusCondition) {
-        this.defaultStatusCondition = defaultStatusCondition;
-    }
-
-    /**
-     * Sets the default date time pattern to use when specifying a date in the email.
-     */
-    public void setDefaultDateTimePattern(String defaultDateTimePattern) {
-        this.defaultDateTimePattern = defaultDateTimePattern;
-    }
-
-    /**
-     * Set the suffix default filename to store the last time an email was sent.
-     * This will be appended to the target and processor ids
-     */
-    public void setDefaultLastDateFilenameSuffix(String defaultLastDateFilenameSuffix) {
-        this.defaultLastDateFilenameSuffix = defaultLastDateFilenameSuffix;
-    }
-
-    /**
-     * Set the directory where the last notification date will be stored.
-     */
-    public void setLastNotificationDateDir(String lastNotificationDateDir) {
-        this.lastNotificationDateDir = lastNotificationDateDir;
-    }
-
-    /**
-     * Sets the {@link EmailFactory} used to generate the emails.
-     */
-    public void setEmailFactory(EmailFactory emailFactory) {
-        this.emailFactory = emailFactory;
-    }
-
-    /**
-     * Sets the {@link ObjectMapper} used to serialize the deployment result.
-     */
-    public void setObjectMapper(final ObjectMapper objectMapper) {
-        this.objectMapper = objectMapper;
-    }
-
-    @Override
-    public void doInit(Configuration config) throws ConfigurationException, DeployerException {
-        templateName = getStringProperty(config, TEMPLATE_NAME_CONFIG_KEY, defaultTemplateName);
-        from = getStringProperty(config, FROM_CONFIG_KEY, defaultFrom);
-        to = getRequiredStringArrayProperty(config, TO_CONFIG_KEY);
-        subject = getStringProperty(config, SUBJECT_CONFIG_KEY, defaultSubject);
-        html = getBooleanProperty(config, HTML_CONFIG_KEY, defaultHtml);
-        serverName = getStringProperty(config, SERVER_NAME_CONFIG_KEY);
-        statusCondition = StatusCondition.valueOf(getStringProperty(config, STATUS_CONDITION_KEY, defaultStatusCondition));
-
-        String processorsMatchRegex = getStringProperty(config, PROCESSOR_MATCH_PATTERNS_CONFIG_KEY);
-        if (processorsMatchRegex != null) {
-            failedProcessorsPattern = Pattern.compile(processorsMatchRegex);
-        }
-        mutePeriodMinutes = getIntegerProperty(config, MUTE_PERIOD_MINUTES_CONFIG_KEY, 0);
-
-        lastDateFilenameSuffix = getStringProperty(config, LAST_DATETIME_FILE_SUFFIX_CONFIG_KEY, defaultLastDateFilenameSuffix);
-
-        if (StringUtils.isEmpty(serverName)) {
-            try {
-                serverName = InetAddress.getLocalHost().getHostName();
-            } catch (UnknownHostException e) {
-                throw new DeployerException("Unable to retrieve localhost address", e);
-            }
-        }
-
-        String dateTimePattern = getStringProperty(config, DATETIME_PATTERN_CONFIG_KEY, defaultDateTimePattern);
-        dateTimeFormatter = DateTimeFormatter.ofPattern(dateTimePattern);
-    }
-
-    @Override
-    protected void doDestroy() {
-        // Do nothing
-    }
-
-    @Override
-    protected ChangeSet doPostProcess(Deployment deployment, ChangeSet filteredChangeSet,
-                                      ChangeSet originalChangeSet) throws DeployerException {
-        Deployment.Status status = deployment.getStatus();
-        if ((statusCondition == StatusCondition.ON_TOTAL_FAILURE && status != Deployment.Status.FAILURE) ||
-            (statusCondition == StatusCondition.ON_ANY_FAILURE && !hasExecutionsFailures(deployment))) {
-                logger.info("Skipping notification because status '{}' does not match the condition '{}'",
-                            status, statusCondition);
-                return null;
-        }
-        if (!matchFailedProcessors(deployment)) {
-            logger.info("Skipping notification because failed processors do not match the configured patterns");
-            return null;
-        }
-
-        if (!shouldSendNotification()) {
-            logger.info("Skipping notification because the mute period has not expired");
-            return null;
-        }
-
-        Map<String, Object> templateModel = new HashMap<>();
-        templateModel.put(SERVER_NAME_MODEL_KEY, serverName);
-        templateModel.put(TARGET_ID_MODEL_KEY, deployment.getTarget().getId());
-        templateModel.put(START_MODEL_KEY, deployment.getStart().format(dateTimeFormatter));
-        templateModel.put(END_MODEL_KEY, deployment.getEnd().format(dateTimeFormatter));
-        templateModel.put(STATUS_MODEL_KEY, deployment.getStatus());
-
-        List<File> attachments = new ArrayList<>();
-        File tempFileDeployment = null;
-
-        try {
-            tempFileDeployment = File.createTempFile("deployment", ".json");
-            objectMapper.writeValue(tempFileDeployment, deployment);
-            attachments.add(tempFileDeployment);
-
-        } catch (IOException e) {
-            logger.error("Could not write deployment as json", e);
-        }
-
-        templateModel.put(OUTPUT_ATTACHED_MODEL_KEY, !attachments.isEmpty());
-
-        try {
-            Email email;
-
-            if (!attachments.isEmpty()) {
-                email = emailFactory.getEmail(from, to, null, null, subject, templateName, templateModel, html,
-                    attachments.toArray(new File[]{}));
-            } else {
-                email = emailFactory.getEmail(from, to, null, null, subject, templateName, templateModel, html);
-            }
-
-            email.send();
-
-            storeNotificationDate();
-            logger.info("Deployment notification successfully sent to {}", Arrays.toString(to));
-        } catch (Exception e) {
-            throw new DeployerException("Error while sending email with deployment report", e);
-        } finally {
-            if(tempFileDeployment != null) {
-                tempFileDeployment.delete();
-            }
-        }
-
-        return null;
-    }
-
-    private File getLastDateFile() {
-        String filename = "%s-%s%s".formatted(targetId, name, lastDateFilenameSuffix);
-        return new File(lastNotificationDateDir, filename);
-    }
-
-    private void storeNotificationDate() {
-        if (mutePeriodMinutes <= 0) {
-            // No need to store the timestamp if there is no mute period
-            return;
-        }
-        try {
-            FileUtils.write(getLastDateFile(), String.valueOf(System.currentTimeMillis()), "UTF-8", false);
-        } catch (IOException e) {
-            logger.warn("Could not store last notification date", e);
-        }
-    }
-
-    private boolean shouldSendNotification() {
-        File lastDateFile = getLastDateFile();
-        if (mutePeriodMinutes <= 0) {
-            logger.info("Mute period is 0 or negative, sending notification");
-            return true;
-        }
-        if (!lastDateFile.exists()) {
-            logger.info("No last notification date file found, sending notification");
-            return true;
-        }
-        try {
-            String lastDate = FileUtils.readFileToString(lastDateFile, "UTF-8").trim();
-            long lastEmailMillis = Long.parseLong(lastDate);
-            long mutePeriodMillis = mutePeriodMinutes * 60 * 1000L;
-            if (System.currentTimeMillis() - lastEmailMillis < mutePeriodMillis) {
-                logger.info("Mute period has not expired, skipping notification");
-                return false;
-            }
-        } catch (NumberFormatException e) {
-            logger.error("Could not parse last notification date from file '{}'", lastDateFile, e);
-        } catch (IOException e) {
-            logger.error("Could not read last notification date from file '{}'", lastDateFile, e);
-        }
-
-        return true;
-    }
-
-    /**
-     * Indicates if any failed deployment processor matches the configured patterns.
-     */
-    private boolean matchFailedProcessors(final Deployment deployment) {
-        return failedProcessorsPattern == null ||
-                deployment.getProcessorExecutions().stream()
-                        .filter(ex -> ex.getStatus() == Deployment.Status.FAILURE)
-                        .map(ProcessorExecution::getProcessorName)
-                        .map(failedProcessorsPattern::matcher)
-                        .anyMatch(Matcher::matches);
-    }
-
-    protected boolean hasExecutionsFailures(Deployment deployment) {
-        for (ProcessorExecution execution : deployment.getProcessorExecutions()) {
-            if (execution.getStatus() == Deployment.Status.FAILURE) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-}
