+++ src/main/resources/base-target.yaml
@@ -246,7 +246,7 @@ target:
       # If emails should be HTML emails
       html: true
       # For which status should the notifications be sent
-      status: ON_ANY_STATUS
+      status: SUCCESS
       # The directory to keep the last notification datetime files
       lastDateDir: ${deployer.main.targets.config.folderPath}
       # The file to keep track of the last notification email datetime
